+++ src/api/products/insights/insights.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Post, Body, Delete, Query, BadRequestException, NotFoundException } from "@nestjs/common"
-import { FrostlakeService } from "./frostlake.service"
+import { InsightsService } from "./insights.service"
 import { CreateAnalyticsDto } from "./dto/create-analytics.dto"
 import { CreateProjectDto } from "./dto/create-project.dto"
 import { TokenAuthorizer, TokenAuthorizerResponse } from "src/authorization/tokenauthorizer/tokenauthorizer.decorator"
@@ -7,14 +7,14 @@ import { ApiKeyAuthorizer, ApiKeyAuthorizerResponse } from "src/authorization/ap
 import { SearchProjectsDto } from "./dto/search-projects.dto"
 import { viewProjectDto } from "./dto/view-project.dto"
 
-@Controller("products/frostlake")
-export class FrostlakeController {
-  constructor(private readonly frostlakeService: FrostlakeService) { }
+@Controller("products/insights")
+export class InsightsController {
+  constructor(private readonly insightsService: InsightsService) { }
 
   @Post("createproject")
   async createProject(@TokenAuthorizer() uft: TokenAuthorizerResponse, @Body() createProjectDto: CreateProjectDto) {
     try {
-      const project = await this.frostlakeService.createProject(uft.workspaceId, createProjectDto)
+      const project = await this.insightsService.createProject(uft.workspaceId, createProjectDto)
       return { project }
     }
 
@@ -27,7 +27,7 @@ export class FrostlakeController {
   async getProjects(@TokenAuthorizer() uft: TokenAuthorizerResponse, @Body() searchProjectsDto: SearchProjectsDto) {
     try {
       const { searchQuery } = searchProjectsDto
-      const projects = await this.frostlakeService.getProjects(uft.workspaceId, searchQuery)
+      const projects = await this.insightsService.getProjects(uft.workspaceId, searchQuery)
       return { projects }
     }
 
@@ -40,7 +40,7 @@ export class FrostlakeController {
   async viewProject(@TokenAuthorizer() uft: TokenAuthorizerResponse, @Body() viewProjectDto: viewProjectDto) {
     try {
       const { projectId } = viewProjectDto
-      const { project, analytics } = await this.frostlakeService.viewProject(uft.workspaceId, projectId)
+      const { project, analytics } = await this.insightsService.viewProject(uft.workspaceId, projectId)
       return { project, analytics }
     }
 
@@ -52,7 +52,7 @@ export class FrostlakeController {
   @Delete("deleteproject")
   async deleteProject(@TokenAuthorizer() uft: TokenAuthorizerResponse, @Query("projectId") projectId: string) {
     try {
-      await this.frostlakeService.deleteProject(uft.workspaceId, projectId)
+      await this.insightsService.deleteProject(uft.workspaceId, projectId)
       return true
     }
 
@@ -64,7 +64,7 @@ export class FrostlakeController {
   @Post("createanalytics")
   async createAnalytics(@ApiKeyAuthorizer() ufak: ApiKeyAuthorizerResponse, @Body() createAnalyticsDto: CreateAnalyticsDto) {
     try {
-      await this.frostlakeService.createAnalytics(ufak.workspaceId, createAnalyticsDto)
+      await this.insightsService.createAnalytics(ufak.workspaceId, createAnalyticsDto)
       return true
     }