+++ src/api/products/products.module.ts
@@ -1,6 +1,6 @@
 import { Module } from "@nestjs/common"
 import { DatalakeModule } from "./datalake/datalake.module"
-import { FrostlakeModule } from "./frostlake/frostlake.module"
+import { InsightsModule } from "./insights/insights.module"
 import { SwapModule } from "./swap/swap.module"
 import { PayModule } from "./pay/pay.module"
 import { NftstudioModule } from "./nftstudio/nftstudio.module"
@@ -8,7 +8,7 @@ import { HyperedgeModule } from "./hyperedge/hyperedge.module"
 import { LedgerscanModule } from "./ledgerscan/ledgerscan.module"
 
 @Module({
-  imports: [DatalakeModule, FrostlakeModule, SwapModule, PayModule, NftstudioModule, HyperedgeModule, LedgerscanModule],
+  imports: [DatalakeModule, InsightsModule, SwapModule, PayModule, NftstudioModule, HyperedgeModule, LedgerscanModule],
 })
 
 export class ProductsModule { }